WIP: add backbone for Rachel
